extract nav reducer from reducer.js
